+++ src/main/java/technion/ir/se/trainers/IParamTrainer.java
@@ -3,5 +3,5 @@ package technion.ir.se.trainers;
 import java.util.Map;
 
 public interface IParamTrainer {
-	public Map<Integer, Double> train();
+	public Map<Double, Double> train();
 }
